added get multi function
function required extra constant's to allow choosing of search pattern. 
current options:
* =
* LIKE
allows the value to be a sql wildcard if % is at the start and end. 
Maybe consider this use case for normal Get however using LIKE for a single item fetch should probably remain exact and not speculative of a LIKE query, since only the first result is returned.